modified

+++ src/app/core/services/backend.service.ts
@@ -13,12 +13,13 @@ export class BackendService {
   // private apiUrl = window.location.origin;
   // private baseUrl = '/servegateway/rest/bdsa/';
   private refreshUrl = this.apiUrl + '/servegateway/rest/bduser/weixin/user/access_token';
+  firstOverdue = true;
 
   jsonHeaders = new Headers({
     'X-Requested-Token': localStorage.getItem('accessToken'),
     'X-Requested-SystemCode' : 'neo_bdsa',
     'X-Requested-DeviceId':  localStorage.getItem('weiXinDeviceId'),
-    'X-Requested-APICode': 'access_token_api',
+    'X-Requested-APICode': 'access_token_weixin_device',
     'X-Requested-Version': '1.0'
   });
   jsonOption = new RequestOptions({ headers: this.jsonHeaders});
@@ -27,6 +28,8 @@ export class BackendService {
     private http: Http,
     private router: Router,
     private oauth: AuthorizeService) {
+      console.log(localStorage.getItem('accessToken'));
+      console.log(localStorage.getItem('refreshToken'));
   }
 
   getAll(url: string ): Promise<any> {
@@ -36,11 +39,14 @@ export class BackendService {
     return this.http.get(this.baseUrl + url, this.jsonOption)
                .toPromise()
                .then(response => {
-                 if (response.json().code === 50013) {
-                     localStorage.clear();
-                     this.getNewToken();
-                 }
-                 return response.json();
+                  if (!localStorage.getItem('weiXinDeviceId')) {
+                    localStorage.clear();
+                    window.location.reload();
+                  }
+                  if (response.json().code === 50013) {
+                    this.getNewToken();
+                  }
+                  return response.json();
                })
                .catch(this.handleError);
   }
@@ -54,40 +60,43 @@ export class BackendService {
   }
 
   getNewToken() {
-    const state: RouterState = this.router.routerState;
-    const snapshot: RouterStateSnapshot = state.snapshot;
-    let url = snapshot.url;
-
-    let headersObj = {
-      'X-Requested-SystemCode' : 'neo_bdsa',
-      'X-Requested-APICode': 'app_api',
-      'X-Requested-Timestamp': Math.floor(new Date().getTime() / 1000),
-      'X-Requested-Nonce': this.MathRand(),
-      'X-Requested-Version': '1.0'
-    };
-    let headers = new Headers(headersObj);
-    let refreshToken = localStorage.getItem('refreshToken');
-    let obj = Object.assign({}, headersObj, {'refreshToken': refreshToken});
-    let form = this.oauth.normalizeParameters(obj);
-    let result = 'POST' + '&' + this.oauth.percentEncode(this.refreshUrl) + '&' + form;
-    let signature = CryptoJS.HmacSHA1(result, result).toString(CryptoJS.enc.Base64);
-    headers.append('X-Requested-Authorization', signature);
-    headers.append('Content-Type', 'application/x-www-form-urlencoded');
-    let body = 'refreshToken=' + refreshToken;
-    this.http.post(this.refreshUrl, body, { headers: headers })
-           .toPromise()
-           .then(response => {
-             if (response.json().code === 50012) {
-               localStorage.clear();
-             } else {
-               localStorage.setItem('accessToken', response.json().data.accessToken);
-               localStorage.setItem('refreshToken', response.json().data.refreshToken);
-               localStorage.setItem('weiXinDeviceId', response.json().data.weiXinDeviceId);
-               localStorage.setItem('user', JSON.stringify(response.json().data));
-             }
-             window.location.reload();
-           })
-           .catch(this.handleError);
+    if (this.firstOverdue) {
+      this.firstOverdue = false;
+      const state: RouterState = this.router.routerState;
+      const snapshot: RouterStateSnapshot = state.snapshot;
+      let url = snapshot.url;
+      let headersObj = {
+        'X-Requested-SystemCode' : 'neo_bdsa',
+        'X-Requested-APICode': 'app_api',
+        'X-Requested-Timestamp': Math.floor(new Date().getTime() / 1000),
+        'X-Requested-Nonce': this.MathRand(),
+        'X-Requested-Version': '1.0'
+      };
+      let headers = new Headers(headersObj);
+      let refreshToken = localStorage.getItem('refreshToken');
+      let obj = Object.assign({}, headersObj, {'refreshToken': refreshToken});
+      let form = this.oauth.normalizeParameters(obj);
+      let result = 'POST' + '&' + this.oauth.percentEncode(this.refreshUrl) + '&' + form;
+      let signature = CryptoJS.HmacSHA1(result, result).toString(CryptoJS.enc.Base64);
+      headers.append('X-Requested-Authorization', signature);
+      headers.append('Content-Type', 'application/x-www-form-urlencoded');
+      let body = 'refreshToken=' + refreshToken;
+      this.http.post(this.refreshUrl, body, { headers: headers })
+             .toPromise()
+             .then(response => {
+               if (response.json().code === 50012) {
+                 localStorage.clear();
+               } else {
+                 localStorage.setItem('accessToken', response.json().data.accessToken);
+                 localStorage.setItem('refreshToken', response.json().data.refreshToken);
+                 localStorage.setItem('weiXinDeviceId', response.json().data.weiXinDeviceId);
+                 localStorage.setItem('user', JSON.stringify(response.json().data));
+               }
+               this.firstOverdue = true;
+               window.location.reload();
+             })
+             .catch(this.handleError);
+    }
   }
 
   private handleError(error: any): Promise<any> {
